v1.3
Changes and fixes:
- Updated hero: McCree --> Cassidy
- Added heroes: Kiriko, Junker Queen, Lifeweaver, Ramattra, and Sojourn
  - Unfortunately, as best I can tell without reinstalling the game, Blizzard removed unique hero colors, so these new heroes all use the same color
- Updated reference API file
- Fixed issue where GUI would offer a 'Miscellaneous' options menu that no longer existed because it was pulling from a static list. Now made dynamic.

Notes:
Despite saying the previous update would be the last, I couldn't stop myself from tinkering.

In the previous build, I had an issue where <except> was failing to catch a TypeError, even though the error was occurring within the <try> block. I ended up using an <if> statement to prevent it from happening in the first place.

This was bugging me, so I made a test file for some quick experiments and finally understood the issue: I was trying to catch multiple exceptions with an <or> operator (except TypeError or KeyError) instead of putting them in a tuple. TIL.

And of course, since I was already in the code, I couldn't stop myself updating the program to include the heroes that were added since I had last worked on it, and making a few fixes along the way.